+++ Plutonium/Output/include/pu/element/Image.hpp
@@ -19,27 +19,27 @@ namespace pu::element
     class Image : public Element
     {
         public:
-            Image(u32 X, u32 Y, std::string Image);
+            Image(uint32_t X, uint32_t Y, std::string Image);
             ~Image();
-            u32 GetX();
-            void SetX(u32 X);
-            u32 GetY();
-            void SetY(u32 Y);
-            u32 GetWidth();
-            void SetWidth(u32 Width);
-            u32 GetHeight();
-            void SetHeight(u32 Height);
+            uint32_t GetX();
+            void SetX(uint32_t X);
+            uint32_t GetY();
+            void SetY(uint32_t Y);
+            uint32_t GetWidth();
+            void SetWidth(uint32_t Width);
+            uint32_t GetHeight();
+            void SetHeight(uint32_t Height);
             std::string GetImage();
             void SetImage(std::string Image);
             bool IsImageValid();
             void OnRender(render::Renderer *Drawer);
-            void OnInput(u64 Down, u64 Up, u64 Held, bool Touch, bool Focus);
+            void OnInput(uint32_t Down, uint32_t Up, uint32_t Held, bool Touch, bool Focus);
         private:
             std::string img;
             render::NativeTexture ntex;
-            u32 x;
-            u32 y;
-            u32 w;
-            u32 h;
+            uint32_t x;
+            uint32_t y;
+            uint32_t w;
+            uint32_t h;
     };
 }
