+++ Plutonium/Output/include/pu/element/ProgressBar.hpp
@@ -19,15 +19,15 @@ namespace pu::element
     class ProgressBar : public Element
     {
         public:
-            ProgressBar(u32 X, u32 Y, u32 Width, u32 Height, double MaxValue);
-            u32 GetX();
-            void SetX(u32 X);
-            u32 GetY();
-            void SetY(u32 Y);
-            u32 GetWidth();
-            void SetWidth(u32 Width);
-            u32 GetHeight();
-            void SetHeight(u32 Height);
+            ProgressBar(uint32_t X, uint32_t Y, uint32_t Width, uint32_t Height, double MaxValue);
+            uint32_t GetX();
+            void SetX(uint32_t X);
+            uint32_t GetY();
+            void SetY(uint32_t Y);
+            uint32_t GetWidth();
+            void SetWidth(uint32_t Width);
+            uint32_t GetHeight();
+            void SetHeight(uint32_t Height);
             draw::Color GetColor();
             void SetColor(draw::Color Color);
             draw::Color GetProgressColor();
@@ -42,12 +42,12 @@ namespace pu::element
             void ClearProgress();
             bool IsCompleted();
             void OnRender(render::Renderer *Drawer);
-            void OnInput(u64 Down, u64 Up, u64 Held, bool Touch, bool Focus);
+            void OnInput(uint32_t Down, uint32_t Up, uint32_t Held, bool Touch, bool Focus);
         private:
-            u32 x;
-            u32 y;
-            u32 w;
-            u32 h;
+            uint32_t x;
+            uint32_t y;
+            uint32_t w;
+            uint32_t h;
             double val;
             double maxval;
             draw::Color clr;
